Add files via upload
got map to plot. added simple popups including location name. 

I am still working on the retrieving data from the website, but in the mean time I just saved the data in the file as a .js in the variable name stream (similar to how we treated the biketheft data).