filter list by window extent in progress
Been trying to get the points into a single layers so I can query from it based on the window extent. Struggling to get the layer to include the points. 

I think I finally got the layer to work, but now I'm working on the actual window extent filtering list. 

8pm - 11/11/18